requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pandas==2.1.1
openpyxl==3.1.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
celery==5.3.1
redis==5.0.1
folium==0.14.0
geopandas==0.14.0
shapely==2.0.1
icalendar==5.0.6
django-leaflet==0.28.3
chart.js==3.9.1
plotly==5.17.0
reportlab==4.0.4
html2canvas==1.4.1
jspdf==2.5.1
papa-parse==5.4.1pandas
openpyxl
